Optimize project structure
